Added JS to add evaluation at the dashboard
A functional JS to add evaluation has been added. No more dummy data is needed in the teacherDashboard.html.

Changed HTML button linking. 

Made changes to the CSS so the buttons at the evaluation cards get the correct styling applied when the JS is making new evaluations inside the dashboard.