Updated the teacherDashboard
Added possibility to print bar graphs when pushing on the result button in a modal window.

The button for adding evaluations now open a new window where you can write in info according to the evaluation (not fully implemented).

Added functionality to remove evaluations by pushing on the remove button.

Changed name from addEvaluation to dashboard on the JS.