Made the datepicker functional
The datepicker can now choose both start and end date. The addEvaluation() function now stores the values in an array, so it is ready to be sent to the backend.

Changed from two to a single input field when choosing date.

Only slight issue is that there is no max limit for how long the evaluation can be open, but it would get limited at the maxDate at 6 months to choose dates.

Added daterangepicker.css and daterangepicker.js to make the datepicker functional with a local library.