Made it possible to submit and get questions inside the evaluations
Since it is in a early stage, the sliders is not updating and it is always opening the evaluation with id = 1.

Changed some naming in the question.java and mainRepository.java to match better the naming from the frontend.

Refactored the evaluation.js for better overview.

Fixed a bug where scrolling did not work in the modals.