Added functionality to studentPage
Moved more files and changed some names to match the new folder structure.

Updated head-tags.

Updated pagesController.java both for student and teacher for right redirect (Not working properly).

Added function to get courseName, but not implemented to the evaluation card.

Added JS-file to generate the buttons at the bottom card inside the evaluation.

Updated and styled the studentEvaluation and added function to load evaluation + questions