Adds possibility to remove current user without giving user id.
An user can now delete itself by sending a DELETE request to /deleteUser instead of /deleteUser/{currentUserId}. This makes it easier to test the project with tools like postman.
Note that it's still possible to delete a user with the previous method.
The Delete user button on user settings seems to be nonfunctional.